Udacity review fixed

+++ application.py
@@ -4,21 +4,17 @@ from collections import defaultdict
 from functools import wraps
 
 import requests
-from flask import Flask, render_template, request, make_response, session, jsonify, \
-    url_for
-from flask_sqlalchemy import SQLAlchemy
+from flask import render_template, request, make_response, session, \
+    jsonify, url_for
 from google.auth.transport import requests as google_requests
 from google.oauth2 import id_token
 from werkzeug.utils import redirect
 
 from forms import Itemform
-
-app = Flask(__name__)
-app.secret_key = secrets.token_hex(32)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'postgresql://udacity:udacity@localhost/udacity'
-db = SQLAlchemy(app)
-
+from models import Item, Category
 # Get google client_id for this app from json file
+from settings import app, db
+
 with open('client_secrets.json', 'r') as f:
     CLIENT_ID = json.load(f)['web']['client_id']
 
@@ -53,7 +49,8 @@ def login():
     else:
         # Check the state variable for extra security
         if session['state'] != request.args.get('state'):
-            response = make_response(json.dumps('Invalid state parameter.'), 401)
+            response = make_response(json.dumps('Invalid state parameter.'),
+                                     401)
             response.headers['Content-Type'] = 'application/json'
             return response
 
@@ -61,9 +58,10 @@ def login():
         token = request.data
 
         # Request an access token from the google api
-        idinfo = id_token.verify_oauth2_token(token, google_requests.Request(), CLIENT_ID)
-        url = 'https://oauth2.googleapis.com/tokeninfo?id_token=%s' % token.decode(
-            encoding='utf-8')
+        idinfo = id_token.verify_oauth2_token(
+            token, google_requests.Request(), CLIENT_ID)
+        url = 'https://oauth2.googleapis.com/tokeninfo?id_token=%s' % \
+              token.decode(encoding='utf-8')
         result = requests.get(url).json()
 
         # If there was an error in the access token info, abort.
@@ -75,15 +73,15 @@ def login():
         # Verify that the access token is used for the intended user.
         user_google_id = idinfo['sub']
         if result['sub'] != user_google_id:
-            response = make_response(
-                json.dumps("Token's user ID doesn't match given user ID."), 401)
+            response = make_response(json.dumps(
+                "Token's user ID doesn't match given user ID."), 401)
             response.headers['Content-Type'] = 'application/json'
             return response
 
         # Verify that the access token is valid for this app.
         if result['aud'] != CLIENT_ID:
-            response = make_response(
-                json.dumps("Token's client ID does not match app's."), 401)
+            response = make_response(json.dumps(
+                "Token's client ID does not match app's."), 401)
             response.headers['Content-Type'] = 'application/json'
             return response
 
@@ -91,8 +89,8 @@ def login():
         stored_access_token = session.get('access_token')
         stored_user_google_id = session.get('user_google_id')
         if stored_access_token and user_google_id == stored_user_google_id:
-            response = make_response(json.dumps('Current user is already connected.'),
-                                     200)
+            response = make_response(json.dumps(
+                'Current user is already connected.'), 200)
             response.headers['Content-Type'] = 'application/json'
             return response
 
@@ -109,8 +107,6 @@ def login():
 
 @app.route("/")
 def home():
-    from models import Item, Category
-
     items = Item.query.limit(10)
     categories = Category.query.all()
     return render_template("home.html", categories=categories, items=items)
@@ -118,19 +114,22 @@ def home():
 
 @app.route("/catalog.json")
 def catalog_json():
-    from models import Category
-
     data = defaultdict(list)
     categories = Category.query.all()
     for category in categories:
         data['category'].append(category.serialize)
     return jsonify(data)
 
+@app.route("/category/<string:name>.json")
+def category_json(name):
+    data = defaultdict(list)
+    category = Category.query.filter(Category.name == name).first()
+    for item in category.items:
+        data['items'].append(item.serialize)
+    return jsonify(data)
 
 @app.route("/category/<string:category>")
 def category_view(category):
-    from models import Category
-
     category = Category.query.filter(Category.name == category).first()
     items = category.items
     return render_template("category.html", category=category, items=items)
@@ -139,14 +138,13 @@ def category_view(category):
 @app.route("/add-item", methods=["GET", "POST"])
 @login_required
 def add_item():
-    from models import Item, Category
-
     form = Itemform()
     form.cat_id.choices = [(cat.id, cat.name) for cat in Category.query.all()]
     if form.validate_on_submit():
         item_new = Item(cat_id=form.data.get('cat_id'),
                         description=form.data.get('description'),
-                        title=form.data.get('title')
+                        title=form.data.get('title'),
+                        author=session.get('user_google_id')
                         )
 
         db.session.add(item_new)
@@ -160,10 +158,12 @@ def add_item():
            methods=["GET", "POST"])
 @login_required
 def edit_item(category, title):
-    from models import Item, Category
-
     cat = Category.query.filter(Category.name == category).first()
-    item = Item.query.filter(Item.cat_id == cat.id, Item.title == title).first()
+    item = Item.query.filter(
+        Item.cat_id == cat.id, Item.title == title).first()
+
+    if session.get('user_google_id') != item.author:
+        return redirect(url_for('home'))
 
     form = Itemform(request.form, obj=item)
     form.cat_id.choices = [(cat.id, cat.name) for cat in Category.query.all()]
@@ -178,20 +178,22 @@ def edit_item(category, title):
 
 @app.route("/category/<string:category>/<string:title>")
 def item_view(category, title):
-    from models import Item, Category
-
     cat = Category.query.filter(Category.name == category).first()
-    item = Item.query.filter(Item.title == title, Item.cat_id == cat.id).first()
+    item = Item.query.filter(
+        Item.title == title, Item.cat_id == cat.id).first()
     return render_template("item.html", item=item)
 
 
-@app.route("/category/<string:category>/<string:title>/delete", methods=["GET", "POST"])
+@app.route("/category/<string:category>/<string:title>/delete",
+           methods=["GET", "POST"])
 @login_required
 def item_del(category, title):
-    from models import Item, Category
-
     cat = Category.query.filter(Category.name == category).first()
-    item = Item.query.filter(Item.title == title, Item.cat_id == cat.id).first()
+    item = Item.query.filter(
+        Item.title == title, Item.cat_id == cat.id).first()
+    if session.get('user_google_id') != item.author:
+        return redirect(url_for('home'))
+
     if request.method == "POST":
         if int(request.form.get("del")):
             db.session.delete(item)
